update borrow count of trending books

+++ src/main/java/com/example/librarymanagement/dtos/models/BookModel.java
@@ -17,8 +17,8 @@ public class BookModel {
     String isbn;
     String imageUrl;
     String pdfPath;
-    int totalCopies;
-    int availableCopies;
+    long totalCopies;
+    long availableCopies;
     int borrowedCopies;
     String status;
 }
